Test concat with lazy arrays
